leumi init

+++ DataProvider/Providers/Banks/Hapoalim/Dto/HapoalimBalanceResponse.cs
@@ -5,13 +5,13 @@ namespace DataProvider.Providers.Banks.Hapoalim.Dto
     public class HapoalimBalanceResponse
     {
         public Object Methadata { get; set; }
-        public Double CurrentAccountLimitsAmount { get; set; }
-        public Double WithdrawalBalance { get; set; }
-        public Double CurrentBalance { get; set; }
-        public Double CreditLimitUtilizationPercent { get; set; }
+        public Decimal CurrentAccountLimitsAmount { get; set; }
+        public Decimal WithdrawalBalance { get; set; }
+        public Decimal CurrentBalance { get; set; }
+        public Decimal CreditLimitUtilizationPercent { get; set; }
         public int CreditLimitUtilizationExistanceCode { get; set; }
         public Object CreditLimitDescription { get; set; }
-        public Double CreditLimitAmount { get; set; }
+        public Decimal CreditLimitAmount { get; set; }
         public String FormattedCurrentAccountLimitsAmount { get; set; }
         public String FormattedCurrentBalance { get; set; }
         public String FormattedWithdrawalBalance { get; set; }
