+++ DataProvider/Providers/Models/Transactions.cs
@@ -12,8 +12,8 @@ namespace DataProvider.Providers.Models
         public DateTime PaymentDate { get; set; }
         public string Description { get; set; }
         public String ProviderName { get; set; } = String.Empty;
-        public double Amount { get; set; }
-        public double CurrentBalance { get; set; }
+        public Decimal Amount { get; set; }
+        public Decimal CurrentBalance { get; set; }
         
         public string SupplierId { get; set; }
     }
