+++ GoldMountainShared/Models/Insur/MortgageInsurAccountDto.cs
@@ -15,10 +15,10 @@ namespace GoldMountainShared.Models.Insur
         public String PlanName { get; set; } = String.Empty;
         public String PolicyId { get; set; } = String.Empty;
         public PolicyStatus PolicyStatus { get; set; }
-        public Double DepositFee { get; set; } = 0;
-        public Double SavingFee { get; set; } = 0;
-        public Double WorkDisabilityMonthly { get; set; } = 0;
-        public Double WorkDisabilityOneTime { get; set; } = 0;
+        public Decimal DepositFee { get; set; } = 0;
+        public Decimal SavingFee { get; set; } = 0;
+        public Decimal WorkDisabilityMonthly { get; set; } = 0;
+        public Decimal WorkDisabilityOneTime { get; set; } = 0;
         public DateTime PolicyOpeningDate { get; set; } = DateTime.MinValue;
         public DateTime ValidationDate { get; set; } = DateTime.MinValue;
         public List<Coverage> Coverage { get; set; } = new List<Coverage>();
